(Test) Mutualize FakeClassificationLearningEnvironment for ClassificationLearningEnvironmentTest and ClassificationLearningAgentTest.
